Rename Cora\Domain\Graphs to Cora\Domain\Graph
