+++ server/src/Domain/Graph/GraphBuilder.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Cora\Domain\Graphs;
+namespace Cora\Domain\Graph;
 
-use Cora\Domain\Graphs\EdgeInterface as Edge;
-use Cora\Domain\Graphs\GraphInterface as IGraph;
+use Cora\Domain\Graph\EdgeInterface as Edge;
+use Cora\Domain\Graph\GraphInterface as IGraph;
 use Cora\Domain\Petrinet\PetrinetInterface as IPetrinet;
 use Cora\Domain\Petrinet\Transition\Transition;
