+++ server/src/Domain/Graph/VertexMapInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Cora\Domain\Graphs;
+namespace Cora\Domain\Graph;
 
 use Ds\Set;
 use IteratorAggregate;
